編集機能実装に伴った改修

+++ src/main/java/com/example/demo/repository/DiaryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Diary;
 
-public interface DiaryRepository extends JpaRepository<Diary, Long> {
+public interface DiaryRepository extends JpaRepository<Diary, Integer> {
 }
